backend/requirements.txt
deepgram-sdk
websockets
python-dotenv
